<2020년 1월 28일 오후 2:45> 풀

+++ booktime/src/main/java/com/ez/booktime/favorite/model/FavoriteServiceImpl.java
@@ -47,8 +47,9 @@ public class FavoriteServiceImpl implements FavoriteService{
 			int nonUserCartDateLimit = 1;
 			
 			cnt = favoriteDao.deleteCartOverDate(userCartDateLimit);
-			cnt = favoriteDao.deleteCartOverDateByNonUser(nonUserCartDateLimit);
-			
+			if(cnt>0) {
+				cnt = favoriteDao.deleteCartOverDateByNonUser(nonUserCartDateLimit);
+			}
 		}catch (RuntimeException e) {
 			e.printStackTrace();
 			cnt = -1;
@@ -62,19 +63,8 @@ public class FavoriteServiceImpl implements FavoriteService{
 	}
 
 	@Override
-	@Transactional
 	public int updateQty(FavoriteVO vo) {
-		int count = 0;
-		for(FavoriteVO fVo :vo.getVoList()) {
-			try{
-				int cnt = favoriteDao.updateQty(fVo);
-				if(cnt>0) count++;
-			}catch (RuntimeException e) {
-				e.printStackTrace();
-				count = -1;
-			}
-		}
-		return count;
+		return favoriteDao.updateQty(vo);
 	}
 
 	@Override
@@ -110,14 +100,14 @@ public class FavoriteServiceImpl implements FavoriteService{
 		int count = 0;
 		
 		if(noList==null) {
-			FavoriteVO vo = selectOneFavorite(Integer.parseInt(favoriteNoList));
+			FavoriteVO vo = favoriteDao.selectOneFavorite(Integer.parseInt(favoriteNoList));
 			vo.setGroup("CART");
 			
 			int cnt = insertFavorite(vo);
 			if(cnt>0) count++;
 		}else {
 			for(String no : noList) {
-				FavoriteVO vo = selectOneFavorite(Integer.parseInt(no));
+				FavoriteVO vo = favoriteDao.selectOneFavorite(Integer.parseInt(no));
 				vo.setGroup("CART");
 				
 				int cnt = insertFavorite(vo);
@@ -126,10 +116,5 @@ public class FavoriteServiceImpl implements FavoriteService{
 		}
 		
 		return count;
-	}
-
-	@Override
-	public FavoriteVO selectOneFavorite(int favoriteNo) {
-		return favoriteDao.selectOneFavorite(favoriteNo);
 	}
 }
